pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = ""
authors = [
    {name = "Kanha Upadhyay",email = "kanha.upadhyay@sifars.com"}
]
readme = "README.md"
requires-python = ">=3.10,<4.0"
dependencies = [
    "fastapi (>=0.116.1,<0.117.0)",
    "uvicorn (>=0.35.0,<0.36.0)",
    "gunicorn (>=23.0.0,<24.0.0)",
    "sqlalchemy (>=2.0.43,<3.0.0)",
    "alembic (>=1.16.5,<2.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "greenlet (>=3.2.4,<4.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "loguru (>=0.7.3,<0.8.0)",
    "pydantic (>=2.11.7,<3.0.0)",
    "aiosqlite (>=0.21.0,<0.22.0)",
    "aiomysql (>=0.2.0,<0.3.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
]

[project.optional-dependencies]
test = [
    "pytest (>=8.0.0,<9.0.0)",
    "pytest-asyncio (>=0.23.0,<0.24.0)",
    "httpx (>=0.27.0,<0.28.0)",
    "pytest-cov (>=4.1.0,<5.0.0)",
    "faker (>=24.0.0,<25.0.0)",
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
addopts = "-v --strict-markers"
markers = [
    "asyncio: mark test as async",
]
